pyproject.toml
[tool.poetry]
name = "pii-masking"
version = "0.1.0"
description = "Lightweight CPU-based PII masking pipeline and API"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "pii_masking", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
onnxruntime = "^1.18.1"
sudachipy = "^0.6.8"
SudachiDict-core = "^20240409"
pydantic = "^2.4.0"
pydantic-settings = "^2.0.0"
numpy = "^1.24.0"
pyyaml = "^6.0"
google-re2 = "^1.1"
python-dotenv = "^1.0.0"
google-cloud-pubsub = "^2.22.0"
duckdb = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.9.0"
isort = "^5.12.0"
mypy = "^1.5.0"
flake8 = "^6.1.0"
locust = "^2.28.1"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/pii_masking",
    "--cov-report=html",
    "--cov-report=term-missing"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
] 